Make Developer Created AWSSignInProvider possible.
Required changes so that AWSSignInProviders can be created and installed.

Allow any number of AWSSignInProviders to maintain sessions and have them
restarted by interceptApplication didFinishLaunchinWithOptions. Without
this fix, only Google and Facebook sessions are restarted. This behavior
is accomplished by a configuration in Info.plist.  If it is missing
prior behaviour with hardcoded Facebook and Google providers is maintained.
This allows Mobile Hub applications to run with no changes to Info.plist.

The configuration relates class names (used to instantiate the singletons)
with user friendly names (used for NSUserDefaultsKeys and for human
readable provider names. The dictionary is located in
AWS->IdentityManager->Default->SignInProviderClassDictionary
and has a form like this:

<key>SignInProviderKeyDictionary</key>
	<dict>
		<key>AWSCUPIdPSignInProvider</key><string>Cognito Pool</string>
		<key>AWSFacebookSignInProvider</key><string>Facebook</string>
		<key>AWSGoogleSignInProvider</key><string>Google</string>
	</dict>

Additionally this fix:

1) Exposes currentSignInProvider, this is necessary so that apps can
determine which provider is active and call it's methods to get access
to properties needed for developer identities and Cognito (your) user pools
(including the ability to get access to the pool).

2) Creates new AWSIdentityManager method providerKey to return the
"Visible" name of an AWSSignInProvider so that the user may be prompted or
notified of which provider is involved (for login errors,
provider lists, etc.).

3) Expose completeLogin. Once the AWSSignInProvider completes the login,
it must call completeLogin, however that method is private.