AWSFacebookSignInProvider has a bug.
AWSFacebookSignInProvider does not call the completion handler when
loginWithCompletionHandler is called if the user presses "Cancel" or "Done".
The code at that point has a comment: // Login canceled, do nothing

This change adds the needed call to the completion handler passing
an error and no result.  The error is: FBSDKLoginUnknownErrorCode
The user info is: "User Cancelled Login"

This error is then (typically - not by mobile hub) displayed to the user.
This is almost exactly how the AWSGoogleSignInProvider behaves, and is
as it should be.

Without this fix behavior on login is different for google and facebook,
and the session state of AWSIdentityManager cannot be maintained