AWSGoogleSignInProvider crashes on startup sometime, I think it is related to trying to start up when it's token is expired or something, because the crash is intermittent, but when it does crash. it keeps crashing until you clear the device settings in simulator.  (Investigation of this is needed.)
However the problem is bigger than that.

If some AWSSignInProvider has a problem, then the app won't start. AWSSignInProviders are many and varied and hopefully more will be created, but we can be sure that going forward we will encounter problems with some one of them.  These may even be temporary server side problems.  And it is bad that a server side problem on one provider might cause an app to be unusable on ALL providers. For a deployed app, that would mean that until "Digits" (for example) fixed it's server that all of my customers that ever used "Digits" might have crashing apps that they could not fix.

So to make AWSIdentityManager more forgiving of problems in AWSSignInProviders. As a fix to AWSGoogleSignInProvider crashes, this fix will reset NSUserDefaults for active sessions, and the keychain on the next App start, when the previous start did not complete AWSSignInProvider initialization.

What the fix does is to set a NSUserDefaults flag (ProvidersOk) that get set to NO when starting and gets set to YES when the AWSSignInProviders have all completed initialization. If upon App start ProviderOk flag is still NO, then the previous app start failed before initializing providers.  In which case AWSIdentitymanager drops all the provider keys and clears the keychain.  That way, if a provider goes bad, the app will function after one crash as long as the user avoids the bad provider. (Also, the providers get a clean slate with no stored sessions, which fixed the AWSGoogleSignInProvider and may even fix other providers.)

The cost is that the user must sign in again but that is better than the app crashing every time it starts.