Allow merging of identities Allow multiple identity merging and simultaneous authenticated identities
Fix AWSIdentityManager so that it tracks logins in a cache and allows
Cognito identity merging.

Controlled using flags set in Info.plist, this allows the developer to choose
how the identityId's are managed.

If there are no flags in the Info.plist behavior is like that of the
existing library (except for one fix, because currently AWSIdentityManager
if you log in to a second identity, gets an error, gets stuck, and has to
be restarted to get rid of it). This ensures Mobile Hub compatibility with
the helper.

The flags are stored in AWS->IdentityManager->Default and are as follows:
<key>IdentityManager</key>
  <dict>
    <key>Default</key>
      <dict>
        <key>Allow Identity Merging</key><true/>
        <key>Allow Simultaneous Active Accounts</key><true/>

Legal values are
TRUE TRUE   - allow both merging and simultaneous active accounts
FALSE TRUE  - allow simultaneous stacked logins but don't merge
              (never returns a merged logins dictionary)
FALSE FALSE - allow one login at a time, any new login
              logs the previous identity out.
              This fixes a current bug see below.

This setting is not legal
FALSE TRUE - Merging identities requires at least two simultaneous logins

Current AWSIdentityManager bug.
If you login twice, each login is independent, they do not merge,
and when you log out of one, you find yourself back logged in as the other
(different) identityId.

This feels strange to the user, and does not really work (I think
because the credentials are not associated with the new identity on
logout (only on login).  At this point you get a whole bunch of logged
errors, and it does not recover gracefully from this condition.

The errors say "GetCredentialsForIdentity keeps failing. Clearing
identityId did not help. Please check your Amazon Cognito Identity
configuration."

But the problem is not that the configuration is bad, it is that
AWSIdentityManager mis-handles the logout so that the identityId remains
old, while the logins dictionary is new.

This fix no longer allows this condition regardless of settings.
If you are not allowing merging, then logging in one user is made to
log out the other first, invisibly and automatically.